manticore: Fix eh_personality definition
There's a link failure that complains that "rust_eh_personality" is
missing. It seems likely to be a bug in Rust because we use the
"eh_personality" attribute to explicitly define the name to be something
else. Work around the issue by renaming the function to
"rust_eh_personality" and making sure the name is not mangled by the
Rust compiler. There's no harm in this workaround nor was I able to find
a proper fix for this. For example, Redox OS uses the same workaround as
we do.