kernel: Add new printf() implementation
This patch adds a new printf() implementation, which is more compatible
with libc and Linux printf implementations than the previously used
tinyprintf. Main motivation for this patch is having a printf()
implementaiton, which is compatible with ACPICA. Unfortunately, neither
tinyprintf or LK's printf() produce sensible output from ACPICA logging,
which is why I had to write one from scratch.

The main issue with previous implementations is that "%8.8X" type of
format specifiers forget to turn on zero-padding and that they misparse
precision to be part of field width, effectively making the field width
88 characters long.