irq: Add user-defined argument to request_irq()
This adds a user-defined argument, or an opaque pointer, to
request_irq(), which is also passed to the interrupt handler. This
allows device drivers, for example, to pass a pointer to a device
structure, so that the handler is able to perform actual work.

We also need to pass information about the device that generated the
interrupt, for example, for device drivers to be able to perform
meaningful work.