x86/apic: Periodic timer support
This adds periodic timer support. We need something to wake up processes
that block on the wait() system call. A periodic timer interrupt is
sufficient for now. However, we eventually want to turn the timer
interrupt in a variable period one-shot timer to avoid waking up the CPU
needlessly, but that's for (much) later.