arch/x86: Fix system call argument passing
The syscall instruction traps to syscall_entry, which delegates to the
syscall() function. However, the syscall() function wants system call
number as the first argument (%rdi), but it is passed in %rax by user
space. We correctly set it, but fail to correctly set rest of the
arguments.

Fix this up by correctly passing all the six possible arguments that we
support for the syscall instruction.