usr/echod: Use liblinux crt0 to fix stack alignment
The echod program is causing general protection exception on SSE
instructions because the stack pointer is not aligned to 16 bytes. The
problem is that main() pushes a frame pointer on the stack (as per
function prologue rules), but the OS starts executing directly in
main(), and no return address has been pushed on the stack.

Fix the problem by switching to liblinux's crt0 routines, which *call*
main() and ensure that stack remains aligned to 16-bytes.