liblinux: Use getevents() in epoll_wait()
This implements parts of epoll_wait() using getevents(). A process that
calls epoll_wait() to query for events is now woken up and notified of
the events. There's still work to do to turn the events into something
usable to the application. We don't, for example, emit the data, just
the event.