kernel: Add atomic ring buffer API
This patch adds an atomic ring buffer API that is used by the kernel and
a user space process to communicate over shared memory. Unsurprisingly,
the underlying implementation is a single-producer, single-consumer
(SPSC) queue that is bounded, lock-free, and wait-free.

The implementation is written from scratch (so all bugs are my own), but
was heavily inspired by Erik Rigtorp's C++ implementation:

  https://github.com/rigtorp/SPSCQueue