kernel: Add getevents() system call
This patch adds a getevents() system call. The system call returns a
pointer to a memory region that is shared by the kernel and a user space
process to communicate with each over using an atomic ring buffer. The
approach is similar to io_uring or XDP's ring buffer, and allows
communication between kernel and user space with as few system calls as
possible.