drivers/virtio-net: Packet I/O events
Register the virtio-net device driver as an event source by the name of
"/dev/eth" and wire up virtio-net packet receive to notify of packet I/O
events to event listeners.

Note that we omit virtio-net headers from the notifications with the
assumption that user space will be consuming Ethernet frames (L2) across
different devices. Also note that the buffer address we pass over to the
user space is not mapped to process address space. Accesses will,
therefore, page fault for now.