vm: Fix VMAddressSpace.populate() for large pages
Fix the VMAddressSpace.populate() function to check the populated region
size and either allocate a large page, a small page, or return an error
if the region size does not match either of them.

Fixes a crash when user space stack size is increased to 2 MB in a
follow up patch.