process: Make user space stack 2 MB in size
The current 4 KB stack is too small and causes the echo server, for
example, to crash when it attempts to allocate a 64 KB receive buffer on
the stack. As we don't support growing the stack via page faults (and
are not sure if we even want to support that), let's make the stack 2 MB
in size for now.

Fixes #24.