kernel, virtio-net: Map received packet buffers to user space
This maps received packet buffers to user space. The way it works is
that the subscribe() system call invokes a per-device operation, which
does the mapping. User space processes can now receive packets from a
virtio device. The event queue is used for (addr, len) tuple
notification of a new packet, and the packet buffer is now mapped to
user space process.