virtio-net: Use VMSpace::allocate() API
Change virtio-net driver to use the allocate() API that does not require
a fixed address.

As a first step, make sure there's exactly one VirtioNetDevice instance
that's used by the interrupt handler and that's returned to the OS
device manager. This was an existing bug that prevented us from setting
"rx_buffer_addr" in acquire() to be used by the interrupt handler that
emits RX events.

Then, switch from allocate_fixed() to allocate() to eliminate hard-coded
virtual memory addresses from the driver.