kernel, libmanticore: Fix out-of-bounds atomic ring buffer emplace
The atomic ring buffer layout looks as follows:

  +--------------------+--------------------------+
  | Header (192 bytes) | Data (rest of the buffer |
  +-----------------------------------------------+

The caller passes a buffer and its size (usually a page) to
atomic_ring_buffer_new(), which calculates capacity as follows:

  buf_size - sizeof(struct atomic_ring_buffer)

However, this can result in the last entry in the ring buffer being to
small to accommodate an element. For example, with a 24-byte element,
the last entry has only room for 16 bytes, which results in an
out-of-bounds write when atomic_ring_buffer_emplace() is called.

Fix up the problem by passing the element size to the
atomic_ring_buffer_new() function, and adjusting capacity so that the
last entry fits a full element.